+++ src/table/components/TableWrapper.jsx
@@ -9,9 +9,9 @@ import TableBodyWrapper from './TableBodyWrapper';
 import TableHeadWrapper from './TableHeadWrapper';
 import FooterWrapper from './FooterWrapper';
 import PaginationContent from './PaginationContent';
-import useDidUpdateEffect from '../../hooks/use-did-update-effect';
-import useFocusListener from '../../hooks/use-focus-listener';
-import useScrollListener from '../../hooks/use-scroll-lietener';
+import useDidUpdateEffect from '../hooks/use-did-update-effect';
+import useFocusListener from '../hooks/use-focus-listener';
+import useScrollListener from '../hooks/use-scroll-lietener';
 import { useContextSelector, TableContext } from '../context';
 import { handleTableWrapperKeyDown } from '../utils/handle-key-press';
 import { updateFocus, handleResetFocus } from '../utils/handle-accessibility';
